% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommimb.R
\name{recommimb}
\alias{recommimb}
\alias{recommimb.default}
\alias{recommimb.formula}
\title{Recommendation of Techniques for Imbalanced Datasets}
\usage{
recommimb(...)

\method{recommimb}{default}(x, y, top.list = 7, ...)

\method{recommimb}{formula}(formula, data, top.list = 7, ...)
}
\arguments{
\item{...}{Further arguments passed to the summarization functions.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{top.list}{Number of recommended techniques. Default is 7 due to experimental evaluations.}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class.
The details section describes the valid values for this group.}
}
\value{
A list of recommendations to be tested by the user
}
\description{
This is a recommender system of techniques for imbalanced datasets. It recommeds
pre-processing and algorithmic-level techniques. The recommendation is based on a
meta-learning approach using traditional meta-features and meta-features designed
specifically for imbalanced datasets.
}
\details{

}
\examples{
## Recommend techniques using formula
data(arsenic_female_bladder)
recommimb(class ~ ., arsenic_female_bladder)

}
\references{

}
\concept{recommmender system}
